requirements.txt
numpy
pandas
matplotlib
scikit-learn
pulp
pyqt5pu